runCI/executeCI.ps1: fix for golang 1.11.0
When using golang docker images from Dockefiles, we want to be sure
that we're using a specific version. For Go 1.11, for example, the
initial released version is 1.11, but it is tagged as 1.11.0 (while
the 1.11 tag refers to whatever is latest 1.11.x version). So, we're
using 1.11.0.

For Dockerfile.windows, though, we can't use 1.11.0 (as there is no such
version available from golang.org) so we're using 1.11 to refer to the
initial release.

Now, there's a version mismatch between the two. To fix, remove the
trailing .0 from the string obtained from Dockerfile.

While at it, make the Select-String patterns a bit more strict by
prepending the pattern with ^ (as we are looking for a string at the
beginning of a line). Otherwise any "ENV GO_VERSION" mentioned in a
comment might result in a false match.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>